+++ CHANGES.md
@@ -1,3 +1,11 @@
+#### [Version 0.1.0 Beta](https://github.com/ForallFramework/loader.package/tree/0.1.0-beta)
+_14-Aug-2013_
+
+* Implemented Event class.
+* Implemented EventDispatcherInterface class.
+* Implemented EventDispatcherTraits class.
+* Implemented the Events core class.
+
 #### [Version 0.0.2 Alpha](https://github.com/ForallFramework/loader.package/tree/0.0.2-alpha)
 _4-June-2013_